+++ tests/unit/test_xml_mapper.py
@@ -5,50 +5,25 @@ import xml.etree.ElementTree as ET
 import pytest
 
 import xmigrate
-
-
-def _make_mapper() -> xmigrate.XMLMapper:
-    """
-    Create an XMLMapper instance with the given source and destination project information.
-
-    Returns
-    -------
-        An instance of XMLMapper initialised with the given project information.
-
-    """
-    source = xmigrate.ProjectInfo(
-        archive_path="/archive/src",
-        id="src_proj",
-        project_name="Source Project",
-        rsync_path="/rsync/src",
-        secondary_id="src_secondary",
-    )
-    destination = xmigrate.ProjectInfo(
-        archive_path="/archive/dst",
-        id="dst_proj",
-        project_name="Destination Project",
-        rsync_path="/rsync/dst",
-        secondary_id="dst_secondary",
-    )
-    return xmigrate.XMLMapper(source=source, destination=destination)
+from tests._helper_functions import make_mapper
 
 
 def test_get_destination_id_returns_mapped_id() -> None:
     """Return the destination ID when a mapping exists."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     mapper.id_map[xmigrate.XnatType.subject]["src_sub_1"] = "dst_sub_1"
     assert mapper.get_destination_id("src_sub_1", xmigrate.XnatType.subject) == "dst_sub_1"
 
 
 def test_get_destination_id_returns_none_for_unknown() -> None:
     """Return None when no mapping exists for the source ID."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     assert mapper.get_destination_id("unknown_id", xmigrate.XnatType.subject) is None
 
 
 def test_rewrite_uris_replaces_source_path() -> None:
     """URI is rewritten from source to destination path."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("file", {"URI": "/archive/src/src_proj/file.dcm"})
     mapper.rewrite_uris(element, "/archive/src/src_proj", "/archive/dst/dst_proj")
     assert element.attrib["URI"] == "/archive/dst/dst_proj/file.dcm"
@@ -56,7 +31,7 @@ def test_rewrite_uris_replaces_source_path() -> None:
 
 def test_rewrite_uris_does_nothing_without_uri() -> None:
     """Element without URI attribute is left unchanged."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("file", {"other": "value"})
     mapper.rewrite_uris(element, "/archive/src/src_proj", "/archive/dst/dst_proj")
     assert "URI" not in element.attrib
@@ -64,7 +39,7 @@ def test_rewrite_uris_does_nothing_without_uri() -> None:
 
 def test_rewrite_uris_raises_if_source_path_missing() -> None:
     """ValueError raised when source path is not found in the URI."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("file", {"URI": "/archive/other/file.dcm"})
     with pytest.raises(ValueError, match="not found in URI"):
         mapper.rewrite_uris(element, "/archive/src/src_proj", "/archive/dst/dst_proj")
@@ -72,7 +47,7 @@ def test_rewrite_uris_raises_if_source_path_missing() -> None:
 
 def test_rewrite_uris_only_replaces_first_occurrence() -> None:
     """Only the first occurrence of source path is replaced."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("file", {"URI": "/archive/src/src_proj/archive/src/src_proj/file.dcm"})
     mapper.rewrite_uris(element, "/archive/src/src_proj", "/archive/dst/dst_proj")
     assert element.attrib["URI"] == "/archive/dst/dst_proj/archive/src/src_proj/file.dcm"
@@ -80,14 +55,14 @@ def test_rewrite_uris_only_replaces_first_occurrence() -> None:
 
 def test_update_id_map_stores_string_id() -> None:
     """ID map stores string representation of destination ID."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     mapper.update_id_map("src_sub_1", "dst_sub_1", xmigrate.XnatType.subject)
     assert mapper.id_map[xmigrate.XnatType.subject]["src_sub_1"] == "dst_sub_1"
 
 
 def test_update_id_map_uses_id_attribute_if_present() -> None:
     """If destination has an .id attribute, that value is stored."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
 
     class FakeListing:
         id = "dst_sub_from_listing"
@@ -98,7 +73,7 @@ def test_update_id_map_uses_id_attribute_if_present() -> None:
 
 def test_map_xml_project_updates_id_and_secondary_id() -> None:
     """Project ID and secondary_ID are updated to destination values."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Project", {"ID": "src_proj", "secondary_ID": "src_secondary"})
     result = mapper.map_xml(element, xmigrate.XnatType.project)
     assert result.attrib["ID"] == "dst_proj"
@@ -107,7 +82,7 @@ def test_map_xml_project_updates_id_and_secondary_id() -> None:
 
 def test_map_xml_project_updates_name() -> None:
     """Project name element text is updated to destination project name."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Project", {"ID": "src_proj", "secondary_ID": "src_secondary"})
     name = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}name")
     name.text = "Source Project"
@@ -117,7 +92,7 @@ def test_map_xml_project_updates_name() -> None:
 
 def test_map_xml_deletes_unwanted_tags() -> None:
     """Tags in tags_to_delete are removed from the element."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Subject", {"project": "src_proj"})
     ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}experiments")
     ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}sharing")
@@ -128,7 +103,7 @@ def test_map_xml_deletes_unwanted_tags() -> None:
 
 def test_map_xml_updates_project_attribute() -> None:
     """The project attribute is updated to the destination project ID."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Subject", {"project": "src_proj", "ID": "sub_1"})
     mapper.map_xml(element, xmigrate.XnatType.subject)
     assert element.attrib["project"] == "dst_proj"
@@ -136,7 +111,7 @@ def test_map_xml_updates_project_attribute() -> None:
 
 def test_map_xml_deletes_id_for_non_project_non_scan() -> None:
     """ID attribute is deleted for types that are not project or scan."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Subject", {"ID": "sub_1", "project": "src_proj"})
     mapper.map_xml(element, xmigrate.XnatType.subject)
     assert "ID" not in element.attrib
@@ -144,7 +119,7 @@ def test_map_xml_deletes_id_for_non_project_non_scan() -> None:
 
 def test_map_xml_preserves_id_for_scan() -> None:
     """ID attribute is preserved for scan resources."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Scan", {"ID": "scan_1", "project": "src_proj"})
     mapper.map_xml(element, xmigrate.XnatType.scan)
     assert element.attrib["ID"] == "scan_1"
@@ -152,7 +127,7 @@ def test_map_xml_preserves_id_for_scan() -> None:
 
 def test_map_xml_remaps_subject_id_tag() -> None:
     """subject_ID tag text is remapped using the id_map."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     mapper.id_map[xmigrate.XnatType.subject]["src_sub_1"] = "dst_sub_1"
     element = ET.Element("Experiment", {"project": "src_proj", "ID": "exp_1"})
     subject_id = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}subject_ID")
@@ -163,7 +138,7 @@ def test_map_xml_remaps_subject_id_tag() -> None:
 
 def test_map_xml_raises_if_tag_remap_missing() -> None:
     """ValueError raised when a tag value has no mapping in id_map."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Experiment", {"project": "src_proj", "ID": "exp_1"})
     subject_id = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}subject_ID")
     subject_id.text = "unmapped_id"
@@ -173,7 +148,7 @@ def test_map_xml_raises_if_tag_remap_missing() -> None:
 
 def test_map_xml_rewrites_file_uris() -> None:
     """File tag URIs are rewritten from source to destination path."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element("Resource", {"project": "src_proj"})
     file_elem = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}file", {"URI": "/archive/src/src_proj/file.dcm"})
     mapper.map_xml(element, xmigrate.XnatType.resource)
@@ -182,7 +157,7 @@ def test_map_xml_rewrites_file_uris() -> None:
 
 def test_map_xml_fixes_mr_scan_modality() -> None:
     """The imageScanData tag is replaced with MRScan tag for MR modality."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element(f"{{{xmigrate.XnatNS.xnat}}}imageScanData", {"ID": "scan_1", "project": "src_proj"})
     modality = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}modality")
     modality.text = "MR"
@@ -192,7 +167,7 @@ def test_map_xml_fixes_mr_scan_modality() -> None:
 
 def test_map_xml_uses_other_scan_for_unknown_modality() -> None:
     """The imageScanData tag is replaced with OtherDicomScan for unknown modality."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element(f"{{{xmigrate.XnatNS.xnat}}}imageScanData", {"ID": "scan_1", "project": "src_proj"})
     modality = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}modality")
     modality.text = "XZ"
@@ -202,7 +177,7 @@ def test_map_xml_uses_other_scan_for_unknown_modality() -> None:
 
 def test_map_xml_uses_other_scan_for_multiple_modalities() -> None:
     """The imageScanData tag is replaced with OtherDicomScan when multiple modalities present."""
-    mapper = _make_mapper()
+    mapper = make_mapper()
     element = ET.Element(f"{{{xmigrate.XnatNS.xnat}}}imageScanData", {"ID": "scan_1", "project": "src_proj"})
     for mod in ("MR", "CT"):
         m = ET.SubElement(element, f"{{{xmigrate.XnatNS.xnat}}}modality")
